quicksave/app/requirements.txt
Quart
Quart-Session
Jinja2
Werkzeug
python-dotenv
hypercorn